% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{AICDiag}
\alias{AICDiag}
\title{AIC and log-likelihood}
\usage{
AICDiag(v, lagv, Theta, Sigma)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}

\item{Theta}{estimated A}

\item{Sigma}{estimated Sigma}
}
\description{
AIC and log-likelihood
}
