% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{AICSca}
\alias{AICSca}
\title{Model selection by AIC}
\usage{
AICSca(v, lagv, theta, sigma)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}

\item{theta}{estimated alpha}

\item{sigma}{estimated sigma}
}
\value{
AIC values and log-likelihood
}
\description{
Model selection by AIC
}
