% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMLE.R
\name{AICSca4for}
\alias{AICSca4for}
\title{AIC and log-likelihood values}
\usage{
AICSca4for(q, omega, LAGVEC, theta, sigma, Mean = NULL)
}
\arguments{
\item{q}{A numerical value with lags}

\item{omega}{A vector of weights}

\item{LAGVEC}{Lagged coordinate vectors from the outputs of \code{\link[=euc.coordvec_lags]{euc.coordvec_lags()}}, \code{\link[=spd.coordvec_lags]{spd.coordvec_lags()}}, and \code{\link[=cor.coordvec_lags]{cor.coordvec_lags()}}}

\item{theta}{Estimated regression coefficients}

\item{sigma}{Estimated s.d.}

\item{Mean}{Logical value with default \code{FALSE} that the regression model does not consider mean reversion term.}
}
\value{
AIC and log-likelihood values
}
\description{
AIC and log-likelihood values
}
