% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thesis.R
\name{anti_combinelist}
\alias{anti_combinelist}
\title{Data structure in thesis}
\usage{
anti_combinelist(Dat)
}
\arguments{
\item{Dat}{List data}
}
\value{
List data for each patients
}
\description{
Data structure in thesis
}
