% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array2vec}
\alias{array2vec}
\title{Array object to vector object}
\usage{
array2vec(Dat, Diag = TRUE)
}
\arguments{
\item{Dat}{An \eqn{p \times p \times n} array object.}

\item{Diag}{Logical value with default value of \code{TRUE}, which means the diagonal entries are included. Otherwise, diagonals are excluded.}
}
\value{
A list data contains \eqn{n} vectors. Dimension of vectors is \eqn{\frac{1}{2}p(p-1)} when \code{Diag = FLASE}, and \eqn{\frac{1}{2}p(p+1)} when \code{Diag = TRUE}.
The vectors are the lower triangular entries in the  symmetric matrix.
}
\description{
Transform array object into vector object.
}
\examples{
### Covariance matrix
S = array(0,c(5,5,10))
for(i in 1:10){S[,,i] = CovM(5)}
array2vec(S,Diag = TRUE)
### Correlation matrix
S = array(0,c(5,5,10))
for(i in 1:10){S[,,i] = CorrM(5)}
array2vec(S,Diag = FALSE)
}
