% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmodel.R
\name{asympCOV}
\alias{asympCOV}
\title{Asymptotic covariance of maximum log-likelihood estimator}
\usage{
asympCOV(sigma, lagv)
}
\arguments{
\item{sigma}{Estimated sigma with respect to noise term}

\item{lagv}{Lagged vectors}
}
\value{
Asymptotic covariance
}
\description{
Its is computed by the Hessian matrix operator and solved by
the inverse of Fisher information matrix.
}
