% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utils.R
\name{auto.mfrow}
\alias{auto.mfrow}
\title{Automatically setup for grid of plots}
\usage{
auto.mfrow(nplots)
}
\arguments{
\item{nplots}{No. of plots. Maximum number is 64}
}
\description{
Automatically setup for grid of plots
}
