% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmodel.R
\name{CalCIs}
\alias{CalCIs}
\title{Compute 95\% confidence interval w.r.t. VAR model with scalar coefficients}
\usage{
CalCIs(alpha, sigma, lagv)
}
\arguments{
\item{alpha}{Estimated alpha}

\item{sigma}{Estimated sigma}

\item{lagv}{lagged vectors}
}
\value{
Confidence matrix with \eqn{m \times 2} where \eqn{m} is the sum of coefficients of
}
\description{
Compute 95\% confidence interval w.r.t. VAR model with scalar coefficients
}
