% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeskySpace.R
\name{cho.Exp}
\alias{cho.Exp}
\title{Exponential map on the Cholesky Space}
\usage{
cho.Exp(L, X)
}
\arguments{
\item{L}{lower triangular matrix with positive diagonal}

\item{X}{lower triangular matrix}
}
\value{
Lower triangular matrix after exponential map
}
\description{
Compute expoenential map on the Cholesky Space emanating from the lower triangular
matrix with positive diagonal \eqn{L} towards the lower triangular matrix \eqn{X}
}
