% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeskySpace.R
\name{cho.geodesic}
\alias{cho.geodesic}
\title{Geodesic on the Cholesky Space}
\usage{
cho.geodesic(t, L1, L2)
}
\arguments{
\item{t}{time \eqn{t, t \in [0,1]}}

\item{L1}{lower triangular matrix}

\item{L2}{lower triangular matrix}
}
\value{
Geodesic
}
\description{
Geodesic emanating from p with direction X
}
