% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeskySpace.R
\name{cho.inner_product}
\alias{cho.inner_product}
\title{Inner product on the Cholesky Space}
\usage{
cho.inner_product(X, Y, L)
}
\arguments{
\item{X}{Tangent vector which is lower triangular matrix  with arbitrary diagonal}

\item{Y}{Tangent vector which is lower triangular matrix  with arbitrary diagonal}

\item{L}{Lower triangular matrix with positive diagonal}
}
\value{
Inner product
}
\description{
Inner product on the Cholesky Space
}
