% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeskySpace.R
\name{cho.Log}
\alias{cho.Log}
\title{Logarithmic map on the CholeskySpace}
\usage{
cho.Log(L1, L2)
}
\arguments{
\item{L1}{lower triangular matrix with positive diagonal}

\item{L2}{lower triangular matrix with positive diagonal}
}
\value{
Lower triangular matrix after logarithm map
}
\description{
Compute the logarithmic map on the Cholesky Space for  the geodesic emanating from
the lower triangular matrix with positive diagonal \eqn{L_1} towards \eqn{L_2}.
}
