% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeskySpace.R
\name{cho.metric}
\alias{cho.metric}
\title{Geodesic distance on the Cholesky Space between two matrices \eqn{L_1}, \eqn{L_2}}
\usage{
cho.metric(L1, L2)
}
\arguments{
\item{L1}{Lower triangular with positive diagonal}

\item{L2}{Lower triangular with positive diagonal}
}
\value{
Riemannian distance
}
\description{
Riemannian distance on the Cholesky Space between two matrices \eqn{L_1} and \eqn{L_2}
that are lower triangular with positive diagonal
}
