% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeskySpace.R
\name{cho.ortho_basis}
\alias{cho.ortho_basis}
\title{Orthonormal basis at \eqn{I_p} on the Cholesky Space}
\usage{
cho.ortho_basis(p)
}
\arguments{
\item{p}{dimensions}
}
\value{
Orthomormal basis and its dimension is \eqn{m = p(p+1)/2}
}
\description{
Orthonormal basis at \eqn{I_p} on the Cholesky Space
}
