% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CholeskySpace.R
\name{cho.para_tran}
\alias{cho.para_tran}
\title{Parallel transport on the Cholesky Space}
\usage{
cho.para_tran(L1, L2, X)
}
\arguments{
\item{L1}{Lower triangular with positive diagonal}

\item{L2}{Lower triangular with positive diagonal}

\item{X}{Tangent vector which is lower triangular matrix  with arbitrary diagonal}
}
\value{
Transported vector \eqn{X} along the geodesic
}
\description{
Parallel transport the tangent vector \eqn{X} at \eqn{L_1} along the geodesic to \eqn{L_2} on
Cholesky Space
}
