% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.coor_vec}
\alias{cor.coor_vec}
\title{Coordinate vector by vectorizing projection in quotient geometry}
\usage{
cor.coor_vec(X, C, D = NULL)
}
\arguments{
\item{X}{A tangent vector matrix in \eqn{T_{C}\mathcal{C}_p^+}.}

\item{C}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{D}{A diagonal matrix in group action \eqn{\mathcal{D}_p^+}. It makes \eqn{I_p, DCD \in \mathcal{S}_p^+} being in optimal positions.
When it is \code{NULL}, we can obtain it from \code{\link[=D.optimal.position]{D.optimal.position()}}.}
}
\value{
A coordinate vector in \eqn{\mathbb{R}^m}
}
\description{
Vectorize a tangent vector matrix \eqn{X \in T_C\mathcal{S}_p^+} into a coordinate vector \eqn{v \in \mathbb{R}^m}. Thereby,
the dimension of affine invariant geometry \eqn{\mathcal{C}_p^+} is \eqn{m = p(p-1)/2}.
}
\examples{
C1 = CorrM(5) ;C2 = CorrM(5)
X1 = cor.LogMap(C1,C2)
cor.coor_vec(X1,C1)
}
