% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{cor.coordvec_lags}
\alias{cor.coordvec_lags}
\title{Coordinate vectors with lags in quotient geometry}
\usage{
cor.coordvec_lags(C, FSM = NULL, maxlag = NULL)
}
\arguments{
\item{C}{An (\eqn{p \times p \times n}) array data consisting of correlation matrices in \eqn{\mathcal{C}_p^+}.}

\item{FSM}{A correlation matrix. It could be the Frechet sample mean S or an attract point.
If this value is missing, the mean reversion coordinate vector will not be considered in the model.}

\item{maxlag}{Maximum lags. The default value is 10.}
}
\value{
A set of coordinate vectors, including lagged vectors, observed vectors,
mean reversion vectors (If FSM is not \code{NULL}), and original matrices.
}
\description{
Compute coordinate vectors in  \eqn{\mathbb{R}^{p(p-1)/2}} including observed tangent vectors,
mean reversion vectors and lagged vectors in quotient geometry \eqn{\mathcal{C}_p^+}.
}
\seealso{
\code{\link[=cor.meanvecs]{cor.meanvecs()}},\code{\link[=cor.paraX]{cor.paraX()}},\code{\link[=cor.lagvecs]{cor.lagvecs()}}
}
