% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSplot.R
\name{cor.dist}
\alias{cor.dist}
\title{Distance matrix using quotient metric}
\usage{
cor.dist(S, parallel = NULL, ncores = NULL)
}
\arguments{
\item{S}{An \eqn{p \times p \times n} array data consisting of full rank correlation matrices.}

\item{parallel}{Logical value with default \code{FALSE}.When it is \code{TRUE},
we use parallel computation to compute distance matrix.}

\item{ncores}{The number of cores used in the parallel computation.}
}
\value{
Distance object as \code{\link[=dist]{dist()}}.
}
\description{
Compute distance matrix of full rank correlation matrices in quotient geometry \eqn{\mathcal{C}_p^+}.
}
\examples{
S = lapply(1:10, function(i) CorrM(5))
S = list2array(S)
cor.dist(S)
}
\seealso{
\code{\link[=cor.metric]{cor.metric()}}
}
