% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.ExpMap}
\alias{cor.ExpMap}
\title{Exponential map in quotient geometry}
\usage{
cor.ExpMap(C, X)
}
\arguments{
\item{C}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{X}{A tangent vector matrix in \eqn{T_{C}\mathcal{C}_p^+}.}
}
\value{
A correlation matrix. It is the mapped point of \eqn{X} at base point \eqn{C} in the manifold \eqn{\mathcal{C}_p^+}.
}
\description{
Exponential map in quotient geometry \eqn{\mathcal{C}_p^+}.
It projects the tangent vector matrix \eqn{X \in T_{C}\mathcal{C}_p^+} at point \eqn{C \in \mathcal{C}_p^+} back to \eqn{\mathcal{C}_p^+}.
}
\details{
For any point \eqn{C} in quotient manifold \eqn{\mathcal{C}_p^+} and tangent vector \eqn{X \in T_C\mathcal{C}_p^+} at the base point \eqn{C},
we can "lift" \eqn{X} from \eqn{T_C\mathcal{C}_p^+} to horizontal space \eqn{\mathcal{H}_C} and then do exponential map in \eqn{\mathcal{S}_p^+} .
Finally, Riemannian submersion is used to project this point onto quotient manifold so that exponential map in \eqn{T_C\mathcal{C}_p^+} is defined as
\deqn{\mathrm{Exp}_{C}(X) = \pi(\mathrm{Exp}_{C}(X^{\sharp})) = \pi (C^{1/2}\exp (C^{-1/2}hor(X) C^{-1/2})C^{1/2}) }
}
\note{
It is the inverse map of \code{\link[=cor.LogMap]{cor.LogMap()}}
}
\examples{
C1 = CorrM(5); C2 = CorrM(5)
X1 = cor.LogMap(C1,C2)
cor.ExpMap(C1,X1)
}
