% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.fsv}
\alias{cor.fsv}
\title{Frechet variance in quotient geometry}
\usage{
cor.fsv(C, FM, parallel = NULL, ncore = NULL)
}
\arguments{
\item{C}{Array object with \eqn{p \times p \times n}.}

\item{FM}{Frechet sample mean. It is obtained from \code{\link[=cor.mean]{cor.mean()}}}

\item{parallel}{Logical values with \code{FALSE} as default value.
If it is \code{TRUE}, we will compute Frechet sample variance by parallel computation, vice versa.}

\item{ncore}{Numer of cores used in parallel computation.}
}
\value{
A numerical value to Frechet function
}
\description{
Compute Frechet sample variance in quotient geometry \eqn{\mathcal{C}_p^+}
}
\examples{
S = lapply(1:10, function(i) CorrM(5))
S = list2array(S)
FM = cor.mean(S,MaxIt = 50)
FM
cor.fsv(S,FM,parallel = TRUE,ncore = 8)
}
