% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.inner_product}
\alias{cor.inner_product}
\title{Inner product in quotient geometry}
\usage{
cor.inner_product(X, Y, S)
}
\arguments{
\item{X}{A tangent vector matrix in \eqn{T_{C}\mathcal{C}_p^+}.}

\item{Y}{A tangent vector matrix in \eqn{T_{C}\mathcal{C}_p^+}.}

\item{S}{A SPD matrix. It lies in the manifold \eqn{\mathcal{S}_p^+} and has horizontal subspace \eqn{\mathcal{H}_S}.
Also, \eqn{C = \pi(S) \in \mathcal{C}_p^+} and \eqn{C \in \mathcal{C}_p^+.}}
}
\value{
A numerical distance.
}
\description{
Compute the inner product for any two tangent vector matrices \eqn{X, Y \in T_C\mathcal{C}_p^+} at the base point \eqn{C \in \mathcal{C}_p^+ }.
}
\details{
For any \eqn{C \in \mathcal{C}_p^+}, \eqn{X,Y \in T_C\mathcal{C}_p^+}, and \eqn{S \in \pi^{-1}(C) \subset \mathcal{S}_p^+},
there exists unique horizontal lift \eqn{X^{\sharp}} and \eqn{Y^{\sharp}}  both in horizontal subspace
\eqn{\mathcal{H}_S} such that \eqn{d\pi_S(X^{\sharp}) = X} and \eqn{d\pi_S(Y^{\sharp}) = Y}.

The quotient metric in quotient manifold \eqn{\mathcal{C}_p^+} is defined based on the affine invariant metric in \eqn{\mathcal{S}_p^+ } as
\deqn{ g^{\mathrm{quo}}_C\langle  X,Y\rangle  = g^{\mathrm{ai}}_S\langle  X^{\sharp},Y^{\sharp} \rangle .}
}
\note{
\itemize{
\item Two tangent vectors \eqn{X,Y} should both lie in the same tangent space \eqn{T_{C}\mathcal{C}_p^+}
\item Quotient metric \eqn{g^{\mathrm{quo}}_C} does not depend on the choice of \eqn{S} in the fiber of \eqn{\pi^{-1}(C) \subset \mathcal{S}_p^+}.
}
}
\examples{
C1 = CorrM(5); C2 = CorrM(5); C3 = CorrM(5)
X = cor.LogMap(C1,C2) ;Y = cor.LogMap(C1,C3)
cor.inner_product(X,Y,C1)
}
