% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{cor.lagvecs}
\alias{cor.lagvecs}
\title{Lagged coordinate vectors from parallel transport in quotient geometry}
\usage{
cor.lagvecs(q, C, paraX, D)
}
\arguments{
\item{q}{A numerical value of lags.}

\item{C}{An (\eqn{p \times p \times n}) array data consisting of correlation matrices in \eqn{\mathcal{C}_p^+}.}

\item{paraX}{Parallel transported tangent vectors from the putput of \code{\link[=cor.paraX]{cor.paraX()}}.}

\item{D}{Diagonal matrices in \eqn{\mathcal{D}_p^+}.
It ensures \eqn{DC_iD} and \eqn{I_p} are in optimal positions in \eqn{\mathcal{S}_p^+}, \eqn{i = 1,\ldots,n}.}
}
\value{
Lagged \eqn{q} coordinate vectors with \eqn{q} lists.
Each list is a matrix with the row of translated coordinate vectors and the column of dimensions of coordinate vectors.
}
\description{
Compute lagged coordinate vectors from parallel transport and arrange the structure of coordinate vectors in quotient geometry \eqn{\mathcal{C}_p^+}
}
\seealso{
\code{\link[=cor.coordvec_lags]{cor.coordvec_lags()}},\code{\link[=cor.paraX]{cor.paraX()}}
}
