% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.LogMap}
\alias{cor.LogMap}
\title{Logarithm maps in quotient geometry}
\usage{
cor.LogMap(C1, C2, D = NULL)
}
\arguments{
\item{C1}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{C2}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{D}{A diagonal matrix in group action \eqn{\mathcal{D}_p^+}. It makes \eqn{C_1, DC_2D \in \mathcal{S}_p^+} being in optimal positions.
When it is \code{NULL}, we can obtain it from \code{\link[=D.optimal.position]{D.optimal.position()}}.}
}
\value{
A tangent vector matrix in \eqn{T_{C_1}\mathcal{C}_p^+}.
}
\description{
Logarithm map in quotient geometry \eqn{\mathcal{C}_p^+}.
It aims to obtain a tangent vector matrix by computing the velocity from point \eqn{C_1} to point \eqn{C_2} both locating in \eqn{\mathcal{C}_p^+}.
}
\details{
Given any \eqn{C_1,C_2 \in \mathcal{C}_p^+}, we can  find the optimal position  in the fibre of \eqn{\pi^{-1}(C_2) \subset \mathcal{S}_p^+}
with respect to \eqn{C_1 \in \mathcal{S}_p^+} by optimizing objective function,
i.e., obtaining \eqn{C_2^* = DC_2D \in \pi^{-1}(C_2)  \subset \mathcal{S}_p^+},
so that \eqn{\mathrm{Log}_{C_1}(C_2^*) \in T_{C_1}\mathcal{C}_p^+} is a tangent vector,
also a horizontal vector, i.e. \eqn{ \mathrm{Log}{C_1}(C_2^*) = hor(\mathrm{Log}_{C_1}(C_2^*))}.
Consequently, we can define the logarithmic map in quotient manifold as
\deqn{\mathrm{Log}_{C_1}(C_2) =  d\pi_{C_1}(hor(\mathrm{Log}_{C_1}(C^*_2))) = d\pi_{C_1}(\mathrm{Log}_{C_1}(C^*_2)) \in T_{C_1}\mathcal{C}_p^+ }
}
\note{
\itemize{
\item The matrix in \eqn{T_{C}\mathcal{C}_p^+} is a symmetric Hollow matrix with vanishing diagonals.
\item The logarithm map  \eqn{X = \mathrm{Log}_{C_1}(C_2)} in quotient manifold is uniquely determined.
\item The tangent vector is the initial velocity at the starting point in the geodesic connecting \eqn{C_1} and \eqn{C_2} both in \eqn{\mathcal{C}_p^+}.
\item It is the inverse map of \code{\link[=cor.ExpMap]{cor.ExpMap()}}.
}
}
\examples{
 C1 = CorrM(5) ; C2 = CorrM(5)
 cor.LogMap(C1,C2)
 D = D.optimal.position(C1,C2)
 cor.LogMap(C1,C2,D)
}
\seealso{
\code{\link[=cor.ExpMap]{cor.ExpMap()}}
}
