% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.mean}
\alias{cor.mean}
\title{Frechet mean in quotient geometry}
\usage{
cor.mean(C, MaxIt, store.M = NULL)
}
\arguments{
\item{C}{Array object with \eqn{p \times p \times n}.}

\item{MaxIt}{Maximum iterations of algorithm. The default value is 150.}

\item{store.M}{Logical values with \code{FALSE} as default value.
If it is \code{TRUE}, we will store the estimated means for each iteration, vice versa.}
}
\value{
Estimated Frechet mean (and each iteration if store.M = \code{TRUE})
}
\description{
Compute Frechet sample mean in quotient geometry \eqn{\mathcal{C}_p^+}
}
\examples{
S = lapply(1:10, function(i) CorrM(5))
S = list2array(S)
FM = cor.mean(S,MaxIt = 50,store.M = TRUE)
FM$mean
}
