% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.mean_Sturm}
\alias{cor.mean_Sturm}
\title{Frechet mean by Sturm algorithm in quotient geometry}
\usage{
cor.mean_Sturm(C, MaxIt, store.est = FALSE, store.fsv = FALSE)
}
\arguments{
\item{C}{Array object with \eqn{p \times p \times n}.}

\item{MaxIt}{Maximum iterations of algorithm.}

\item{store.est}{Logical values with \code{FALSE} as default value. If it is \code{TRUE}, we will store the estimated means for each iteration, vice versa.}

\item{store.fsv}{Logical values with \code{FALSE} as default value. If it is \code{TRUE}, we will store the estimated variance for each iteration, vice versa.}
}
\value{
Frechet mean
}
\description{
Compute Frechet sample mean in quotient geometry \eqn{\mathcal{C}_p^+}.
}
