% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{cor.meanvecs}
\alias{cor.meanvecs}
\title{Coordinate vectors of mean reversion in quotient geometry}
\usage{
cor.meanvecs(C, FSM, D)
}
\arguments{
\item{C}{An (\eqn{p \times p \times n}) array data consisting of correlation matrices in \eqn{\mathcal{C}_p^+}.}

\item{FSM}{A correlation matrix. It could be Frechet sample mean or  a given attract point.}

\item{D}{Diagonal matrices in \eqn{\mathcal{D}_p^+}.
It ensures \eqn{DC_iD} and \eqn{I_p} are in optimal positions in \eqn{\mathcal{S}_p^+}, \eqn{i = 1,\ldots,n}.}
}
\value{
A matrix containing coordinate vectors of mean reversion with the row of observations and the column of the dimension \eqn{m}
}
\description{
Compute coordinate vectors at the centre point or attract point in quotient geometry  \eqn{\mathcal{C}_p^+}.
}
\seealso{
\code{\link[=cor.coordvec_lags]{cor.coordvec_lags()}}
}
