% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.metric}
\alias{cor.metric}
\title{Geodesic distance in  quotient geometry}
\usage{
cor.metric(C1, C2, D = NULL)
}
\arguments{
\item{C1}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{C2}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{D}{A diagonal matrix in group action \eqn{\mathcal{D}_p^+}. It makes \eqn{C_1, DC_2D \in \mathcal{S}_p^+} being in optimal positions.
When it is \code{NULL}, we can obtain it from \code{\link[=D.optimal.position]{D.optimal.position()}}}
}
\value{
A numerical value of distance between \eqn{C_1} and \eqn{C_2} in \eqn{\mathcal{C}_p^+}
}
\description{
Given any two points \eqn{C_1} and \eqn{C_2} in quotient geometry \eqn{\mathcal{C}_p^+},
we compute the geodesic distance between these two points based on affine invariant metric in \eqn{\mathcal{S}_p^+}.
}
\details{
Given any two points \eqn{C_1,C_2 \in \mathcal{C}_p^+} and chosen \eqn{\pi^{-1}(C_1) = I_pC_1I_p}, the quotient metric in quotient manifold is defined a
\deqn{d^{\mathrm{quo}}(C_1,C_2) = \inf \limits_{D \in \mathcal{D}_p^+} d^{\mathrm{ai}}(C_1,DC_2D) }
}
\examples{
C1 = CorrM(5) ; C2 = CorrM(5)
cor.metric(C1,C2)
}
