% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{cor.ortho_basis}
\alias{cor.ortho_basis}
\title{Orthonormal basis at any base point in quotient geometry}
\usage{
cor.ortho_basis(C, D = NULL)
}
\arguments{
\item{C}{A correlation matrix in \eqn{\mathcal{C}_p^+}}

\item{D}{A diagonal matrix in group action \eqn{\mathcal{D}_p^+}. It makes \eqn{C_1, DC_2D \in \mathcal{S}_p^+} being in optimal positions.
When it is \code{NULL}, we can obtain it from \code{\link[=D.optimal.position]{D.optimal.position()}}.}
}
\value{
A list of orthonormal basis in the tangent space \eqn{T_{C}\mathcal{C}_p^+}
containing \eqn{m = p(p - 1)/2} observations of \eqn{p \times p} basis.
}
\description{
Construct orthonormal basis in the tangent space \eqn{T_{C}\mathcal{C}_p^+}.
}
\details{
Let  \eqn{C} be any point in quotient manifold \eqn{\mathcal{C}_p^+}. The orthonormal basis in tangent space at base point \eqn{C} is defined as
\deqn{ \begin{aligned}
 \boldsymbol{E}^C  & = \mathcal{P}_{I_p \rightarrow C}(\boldsymbol{E}^{I_p}) \\
 & =  d \pi_{C^{\dag}}\left[(C^{\dag} I_p^{-1})^{1/2}(\boldsymbol{E}^{I_p})^{\sharp} ((C^{\dag} I_p^{-1})^{1/2})^T \right]\\
 & = d \pi_{C^{\dag}}\left[{C^{\dag}}^{1/2}hor(\boldsymbol{E}) {C}^{1/2} \right]
 \end{aligned}
 }
where \eqn{\boldsymbol{E}^{I_p}} is the orthonormal basis in the tangent space \eqn{T_{I_p}\mathcal{C}_p^+} with the dimension of \eqn{m = p(p-1)/2},
and it is equivalent to orthonormal basis in horizontal space \eqn{\mathcal{H}_{I_p}} in \code{\link[=hor.ortho_basis]{hor.ortho_basis()}}.
And, \eqn{C^{\dag} = DCD} is the optimal position with respect to \eqn{I_p}, both locating in Riemannian manifold \eqn{\mathcal{S}_p^+}.
}
\examples{
C1 = CorrM(5)
cor.ortho_basis(C1)
}
