% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{cor.paraX}
\alias{cor.paraX}
\title{Parallel transport vectors with maximum lags in quotient geometry}
\usage{
cor.paraX(maxlag, C, X)
}
\arguments{
\item{maxlag}{A numerical value which represents the maximum lags involved.}

\item{C}{An (\eqn{p \times p \times n}) array data consisting of correlation matrices in \eqn{\mathcal{C}_p^+}.}

\item{X}{A \eqn{n-1} length list data consisting of tangent vectors.}
}
\value{
The parallel transported vectors.
}
\description{
Compute lagged tagent vector matrix by parallel transport in affine invariant geometry \eqn{\mathcal{C}_p^+} .
}
\details{
Given any original data point \eqn{C} in \eqn{\mathcal{C}_p^+} and corresponding
tangent vectors \eqn{X}, we compute all parallel transport vectors with the maximum lags \eqn{maxlag}.
For example, \code{maxlag = 3}, the parallel transported vector could be written as
\eqn{X_{4,1} = \mathcal{P}_{C_1,C_4}(X_1)}, \eqn{X_{4,2} = \mathcal{P}_{C_2,C_4}(X_2)}, and \eqn{X_{4,3} = \mathcal{P}_{C_3,S_4}(C_3)}.
The basic structure (e.g maxlag = 3) stored in outputs is
\eqn{\mathcal{P}_{C_1,C_2}(X_1)};

\eqn{\mathcal{P}_{C_2,C_3}(X_2)}, \eqn{\mathcal{P}_{C_1,C_3}(X_1)};

\eqn{\mathcal{P}_{C_3,C_4}(X_3)}, \eqn{\mathcal{P}_{C_2,C_4}(X_2)} ,\eqn{\mathcal{P}_{C_1,C_4}(X_1)};

\eqn{\mathcal{P}_{C_4,C_5}(X_4)}, \eqn{\mathcal{P}_{C_3,C_5}(X_3)} ,\eqn{\mathcal{P}_{C_2,C_5}(X_2)}
}
\seealso{
\code{\link[=cor.lagvecs]{cor.lagvecs()}}
}
