% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cor.simulate}
\alias{cor.simulate}
\title{Simulate full rank correlation matrices under the framework of manifold-adapted model}
\usage{
cor.simulate(n, alpha, sigma, beta = NULL, attC = NULL, p = NULL)
}
\arguments{
\item{n}{Number of simulated matrix-valued time points.}

\item{alpha}{Coefficients in the model whose length is the lags of autoregressive terms.
If it is a vector, the manifold-adapted model with scalar coefficients is used.
If it is a list with multiple diagonals, the model with diagonal coefficients is used.}

\item{sigma}{Standard deviation of white noise.}

\item{beta}{Coefficient of mean reversion. It follows the attract point/ Mean.
When we consider the mean-reverting term in the model, \code{attC} should be input and then \code{beta} is required.}

\item{attC}{Attract point.}

\item{p}{Dimensions if \eqn{beta} and \eqn{attC} are missing.}
}
\value{
An array data \eqn{p \times p \times n} in  \eqn{\mathcal{C}_p^+}.
}
\description{
Simulate full rank correlation matrices under framework of manifold-adapted model in \eqn{\mathcal{C}_p^+}.
}
