% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CovM}
\alias{CovM}
\title{Full-rank covariance matrix}
\usage{
CovM(p)
}
\arguments{
\item{p}{Row (or column) number of the symmetric matrix.}
}
\value{
A covariance matrix with full rank. It is a SPD matrix.
}
\description{
A easy way to create an arbitrary \eqn{p \times p} covariance matrix with full rank.
}
\details{
\eqn{M \in \mathbb{R}^{p \times p}}  is a \eqn{p \times p} symmetric matrix
which elements \eqn{M_{i,j},i,j = 1,...,p} are simulated from a normal distribution \eqn{N(0,0.5^2)},
and then full-rank covariance matrix is computed by
\deqn{S = \frac{1}{2}(M+M^T)}
\deqn{S = SS^T}
}
\examples{
CovM(5)

}
\seealso{
\code{\link[=wrapGau]{wrapGau()}}
}
