% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{D.optimal.position}
\alias{D.optimal.position}
\title{Find the optimal position}
\usage{
D.optimal.position(
  C1,
  C2,
  MaxIt = NULL,
  delta = NULL,
  conver = NULL,
  store.D = NULL
)
}
\arguments{
\item{C1}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{C2}{A correlation matrix in \eqn{\mathcal{C}_p^+}.}

\item{MaxIt}{Maximum iterations. Default is 200.}

\item{delta}{Step size. Default is 0.001.}

\item{conver}{Convergence parameter. Default is 4, which means the distance between two iterations is less then \eqn{10^{-4}}.}

\item{store.D}{Logical values. Default is \code{FALSE} which means we don't want to store the iterated \eqn{D} and the values of objective function.}
}
\value{
A diagonal matrix (iterated \eqn{D} and values of objective function if store.M = \code{TRUE}).
}
\description{
Find the optimal position for any two points \eqn{S_1 \in \pi^{-1}(C_1), S_2 \in \pi^{-1}(C_2)}
in \eqn{\mathcal{S}_p^+} via greedy descent algorithm.
}
\details{
Suppose \eqn{C_1} and \eqn{C_2} both in \eqn{\mathcal{C}_p^+},
\eqn{C_1} lies in \eqn{\mathcal{S}_p^+} as it is a SPD matrix.
We aim to find a matrix \eqn{D \in \mathcal{D}_p^+}
which makes affine invariant metric of \eqn{C_1} and \eqn{DC_2D} in \eqn{\mathcal{S}_p^+} shortest.
}
