% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{deri_submer}
\alias{deri_submer}
\title{Derivative of the Riemannian submersion}
\usage{
deri_submer(V, S)
}
\arguments{
\item{V}{A tangent vector matrix in \eqn{T_S\mathcal{S}_p^+}.}

\item{S}{A SPD matrix in \eqn{\mathcal{S}_p^+}}
}
\value{
A tangent vector matrix in the tangent space \eqn{T_{\pi(S)}\mathcal{C}_p^+}.
}
\description{
Differentiate Riemannian submersion \eqn{\pi} of \eqn{V \in T_S\mathcal{S}_p^+} at base point \eqn{S \in \mathcal{S}_p^+}.
}
\details{
The tangent vector matrix \eqn{X} is computed by the derivative of Riemannian submersion \eqn{d\pi_S(V)},
i.e. \eqn{X = d\pi_S(V)} is the  projected vector of \eqn{V} onto the tangent space \eqn{T_{\pi(S)}\mathcal{C}_p^+}.
}
\note{
The tangent vector \eqn{X} by derivative of the Riemannian submersion \eqn{d\pi_S(V)} is a symmetric Hollow matrix with vanishing diagonals.
}
\examples{
S1 = CovM(5)
S2 = CovM(5)
V1 = spd.LogMap(S1,S2)
deri_submer(V1,S1)
}
