% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Diag}
\alias{Diag}
\title{Diagonal matrix}
\usage{
Diag(M)
}
\arguments{
\item{M}{A square matrix.}
}
\value{
A square diagonal matrix.
}
\description{
Preserve diagonals and discard off-diagonal entries of a square matrix \eqn{M}.
}
\examples{
M = matrix(c(1:9),3,3)
Diag(M)
}
