% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diag_ind}
\alias{diag_ind}
\title{Indices of diagonal positions from a square matrix}
\usage{
diag_ind(p)
}
\arguments{
\item{p}{Row (or column) number of the square matrix.}
}
\description{
Extract indices of diagonal positions from a square matrix. We number each position from a square matrix.
For example a \eqn{3 \times 3} matrix, there are indices of entries from 1 to 9. Therefore, the indices of diagonal positions is \eqn{(1,5,9)}.
}
\examples{
diag_ind(3)

}
