% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR.R
\name{DR_GME}
\alias{DR_GME}
\title{Dimensionality reduction via minimizing redundancy}
\usage{
DR_GME(X)
}
\arguments{
\item{X}{An \eqn{p \times p \times n} array data.}
}
\value{
The descending order with choosing pairs and corresponding GMEs
}
\description{
Dimensional reduction for full rank covariance (or correlation) matrices by
choosing the greatest mean of minimum eigenvalues (GME) via greedy algorithm.
}
\details{
Suppose we have a set of channels \eqn{\mathfrak{C}}, and we let \eqn{S_i'(\mathfrak{C})}
(or \eqn{C_i'(\mathfrak{C})}) denote the restriction of the full covariance (or correlation)
matrix \eqn{S_i'} (or \eqn{C'_i}) to \eqn{\mathfrak{C}}. Also define
\deqn{\omega(\mathfrak{C})  = \frac{1}{n}\sum_{i}\min \sigma[S_i'(\mathfrak{C})] }
or \deqn{\omega(\mathfrak{C})  = \frac{1}{n}\sum_{i}\min \sigma[C_i'(\mathfrak{C})]}
where \eqn{\sigma(A)} denotes the set of eigenvalues of \eqn{A \in  \mathrm{Sym}(p)}.
A greedy algorithm was used to construct sets \eqn{\mathfrak{C}} which maximized \eqn{\omega(\mathfrak{C})} for a fixed value of \eqn{p}.
More specifically, we first consider every set \eqn{\mathfrak{C}} consisting of a pair of channels to find the optimal set \eqn{\mathfrak{C}_2}.
We then consider every set \eqn{\mathfrak{C}} containing \eqn{\mathfrak{C}_2}
and one additional channel, to find the optimal set \eqn{\mathfrak{C}_3}, and so on.
}
\examples{
data(EEG)
eegcov = EEG$seizure1
DR_GME(eegcov)
}
