% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR.R
\name{DR_HME}
\alias{DR_HME}
\title{Dimensionality reduction by choosing the highest minimum eigenvalues via greedy algorithm}
\usage{
DR_HME(X)
}
\arguments{
\item{X}{An \eqn{p \times p \times n} array data.}
}
\value{
The descending order of HME with choosing pairs and corresponding HMEs
}
\description{
Dimensionality reduction by choosing the highest minimum eigenvalues via greedy algorithm
}
\examples{
data(EEG)
eegcov = EEG$seizure1
DR_HME(eegcov)
}
