% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EEG}
\alias{EEG}
\title{EEG data for two examples}
\format{
A list data with 4 elements consisting of \eqn{p \times p \times n} covariance matrices
}
\source{
{EEG singnal data}
}
\usage{
EEG
}
\description{
EEG data for two examples
}
\examples{
data(EEG)

}
\keyword{datasets}
