% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstDiagSig}
\alias{EstDiagSig}
\title{Standard deviation in the manifold-adapted model with diagonal coefficients}
\usage{
EstDiagSig(v, lagv, Theta)
}
\arguments{
\item{v}{A \eqn{n \times m} matrix with \eqn{m} dimensions and \eqn{n} observations.}

\item{lagv}{A list data with the length of \eqn{q} containing lagged vectors  that each of it has the same structure with \code{v}.}

\item{Theta}{A list of estimated diagonal regressive coefficients with the length of \eqn{q} from the output of \code{\link[=EstScaTheta]{EstScaTheta()}}.}
}
\value{
A vector of diagonal stimated diagonal of Sigma
}
\description{
Compute standard deviation of white noise in the manifold-adapted model with diagonal coefficients.
}
\seealso{
\code{\link[=EstScaTheta]{EstScaTheta()}}
}
