% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstDiagTheta}
\alias{EstDiagTheta}
\title{Estimate diagonal matrices A in VAR model}
\usage{
EstDiagTheta(v, lagv)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}
}
\value{
estimated diagonals of coefficients
}
\description{
Estimate diagonal matrices A in VAR model
}
