% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstSca}
\alias{EstSca}
\title{Coefficients in the manifold-adapted model with scalar coefficients}
\usage{
EstSca(v, lagv)
}
\arguments{
\item{v}{A \eqn{n \times m} matrix with \eqn{m} dimensions and \eqn{n} observations.}

\item{lagv}{A list data with the length of \eqn{q} containing lagged vectors  that each of it has the same structure with \code{v}.}
}
\value{
Estimated scalar coefficients, including \describe{
\item{theta}{Estimated scalar regressive coefficients.}
\item{sigma}{Estimated standard deviation of the white noise.}
}
}
\description{
Estimate coefficients in the manifold-adapted model with scalar coefficients,
including coefficients in regressive terms and standard deviation of the white noise.
}
