% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstScaSig}
\alias{EstScaSig}
\title{estimate Sigma in scalar model}
\usage{
EstScaSig(v, lagv, theta)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}

\item{theta}{estimated alpha}
}
\value{
estimated sigma
}
\description{
estimate Sigma in scalar model
}
\seealso{
The whole function could find in function \code{\link[=MLE.Sca]{MLE.Sca()}},\code{\link[=EstSca]{EstSca()}},\code{\link[=EstScaTheta]{EstScaTheta()}}
}
