% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMLE.R
\name{EstScaSig4all}
\alias{EstScaSig4all}
\title{Estimated s.d.}
\usage{
EstScaSig4all(q, omega, LAGVEC, theta, Mean = NULL)
}
\arguments{
\item{q}{A numerical value with lags}

\item{omega}{A vector of weights}

\item{LAGVEC}{Lagged coordinate vectors from the outputs of \code{\link[=euc.coordvec_lags]{euc.coordvec_lags()}}, \code{\link[=spd.coordvec_lags]{spd.coordvec_lags()}}, and \code{\link[=cor.coordvec_lags]{cor.coordvec_lags()}}}

\item{theta}{Estimated regression coefficients}

\item{Mean}{Logical value with default \code{FALSE} that the regression model does not consider mean reversion term.}
}
\value{
Estimated s.d.
}
\description{
Estimated s.d.
}
