% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstScaTheta}
\alias{EstScaTheta}
\title{Regressive coefficients in the manifold-adapted model with scalar coefficients}
\usage{
EstScaTheta(v, lagv)
}
\arguments{
\item{v}{A \eqn{n \times m} matrix with \eqn{m} dimensions and \eqn{n} observations.}

\item{lagv}{A list data with the length of \eqn{q} containing lagged vectors that each of it has the same structure with \code{v}.}
}
\value{
A vector of estimated regressive coefficients.
}
\description{
Compute regressive coefficients in the manifold-adapted model with scalar coefficients.
}
