% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMLE.R
\name{EstScaTheta4all}
\alias{EstScaTheta4all}
\title{Estimated regression coefficients}
\usage{
EstScaTheta4all(q, omega, LAGVEC, Mean = NULL)
}
\arguments{
\item{q}{A numerical value with lags}

\item{omega}{A vector of weights}

\item{LAGVEC}{Lagged coordinate vectors from the outputs of \code{\link[=euc.coordvec_lags]{euc.coordvec_lags()}}, \code{\link[=spd.coordvec_lags]{spd.coordvec_lags()}}, and \code{\link[=cor.coordvec_lags]{cor.coordvec_lags()}}}

\item{Mean}{Logical value with default \code{FALSE} that the regression model does not consider mean reversion term.}
}
\value{
Estimated regression coefficients
}
\description{
Estimated regression coefficients
}
