% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.coor_vec}
\alias{euc.coor_vec}
\title{Coordinate vector by vectorizing projection in symmetric matrix space}
\usage{
euc.coor_vec(V, corMat = NULL)
}
\arguments{
\item{V}{A tangent vector matrix.}

\item{corMat}{Logical values with default value \code{FALSE}.
If it is \code{TRUE}, symmetric matrix space is consisting of correlation matrix and the dimension of coordinate vector \eqn{\boldsymbol{v}} is
\eqn{m =p(p-1)/2}. Otherwise, \eqn{m = p(p+1)/2}.}
}
\value{
A coordinate vector in \eqn{\mathbb{R}^m}
}
\description{
Vectorize a tangent vector matrix \eqn{V \in T_S\mathrm{Sym}(p)} into a coordinate vector \eqn{\boldsymbol{v} \in \mathbb{R}^m}. Thereby,
the dimension of symmetric matrix space is \eqn{m = p(p+1)/2}. If the symmetric matrix space consists of correlation matrices,
the dimension of symmetric matrix space will be \eqn{m = p(p-1)/2}.
}
\examples{
S1 = CovM(5)
S2 = CovM(5)
V = euc.LogMap(S1,S2)
euc.coor_vec(V)
}
\seealso{
\code{\link[=euc.Euc2Sym]{euc.Euc2Sym()}}
}
