% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{euc.coordvec_lags}
\alias{euc.coordvec_lags}
\title{Coordinate vectors with lags in symmetric matrix space}
\usage{
euc.coordvec_lags(S, FSM = NULL, maxlag = NULL, corMat = NULL)
}
\arguments{
\item{S}{An \eqn{p \times p \times n} array data consisting of symmetric matrices.}

\item{FSM}{A symmetric matrix. It could be the Frechet sample mean \eqn{S} or an attract point.
If this value is missing, the mean reversion coordinate vector will not be considered in the model.}

\item{maxlag}{Maximum lag. The default value is 10.}

\item{corMat}{Logical values. If it is \code{TRUE}, the dimension of coordinate vector is
\eqn{p(p+1)/2 } , otherwise, \eqn{p(p-1)/2}. The default value is \code{FALSE}.}
}
\value{
A set of coordinate vectors, including lagged vectors, observed vectors,
mean reversion vectors (If FSM is not \code{NULL}), and original matrices.
}
\description{
Compute coordinate vectors in  \eqn{\mathbb{R}^m} including observed tangent vectors,
mean reversion vectors and lagged vectors in symmetric matrix space  \eqn{\mathrm{Sym}(p)}.
}
