% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSplot.R
\name{euc.dist}
\alias{euc.dist}
\title{Distance matrix using Euclidean metric}
\usage{
euc.dist(S)
}
\arguments{
\item{S}{An \eqn{p \times p \times n} array data consisting of symmetric matrices.}
}
\value{
Distance object as \code{\link[=dist]{dist()}}.
}
\description{
Compute distance matrix of symmetric matrices in symmetric matrix space \eqn{\mathrm{Sym}(p)}.
}
\examples{
data(EEG)
redEEG = DR_PCA(EEG$seizure1,N0 = 15)
euc.dist(redEEG)
}
\seealso{
\code{\link[=euc.dist]{euc.dist()}}
}
