% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.Euc2Sym}
\alias{euc.Euc2Sym}
\title{Mapping back the coordinate vector to the data point in symmetric matrix space}
\usage{
euc.Euc2Sym(v, S, corMat = NULL)
}
\arguments{
\item{v}{A coordinate vector in \eqn{\mathbb{R}^m}.}

\item{S}{A symmetric matrix in \eqn{\mathrm{Sym}(p)}.}

\item{corMat}{Logical values with default value \code{FALSE}. If it is \code{TRUE},
symmetric matrix space is consisting of correlation matrix and the dimension of coordinate vector \eqn{\boldsymbol{v}} is
\eqn{m =p(p-1)/2}. Otherwise, \eqn{m = p(p+1)/2}.}
}
\value{
A symmetric matrix in \eqn{\mathrm{Sym}(p)}
}
\description{
Map back the coordinate vector to the tangent space \eqn{T_{I_p}\mathrm{Sym}(p)}
and then do exponential map at base point \eqn{S \in \mathrm{Sym}(p)}.
}
\examples{
S1 = CovM(5)
S2 = CovM(5)
v = euc.coor_vec(S2-S1)
euc.Euc2Sym(v,S1)

# Correlation matrix
C1 = CorrM(5);C2 = CorrM(5)
u = euc.coor_vec(C2-C1,corMat = TRUE)
euc.Euc2Sym(u,C1,corMat = TRUE)
}
\seealso{
\code{\link[=euc.coor_vec]{euc.coor_vec()}}
}
