% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.ExpMap}
\alias{euc.ExpMap}
\title{Exponential map in symmetric matrix space}
\usage{
euc.ExpMap(S, V)
}
\arguments{
\item{S}{A symmetric matrix.}

\item{V}{A tangent vector matrix.}
}
\value{
A symmetric matrix. It is the mapped point of \eqn{V} at base point \eqn{S} in symmetric matrix space \eqn{\mathrm{Sym}(p)}.
}
\description{
Exponential map in symmetric matrix space \eqn{\mathrm{Sym}(p)}.
It projects tangent vector matrix \eqn{V \in T_{S}\mathrm{Sym}(p)} at point \eqn{S \in \mathrm{Sym}(p)} back to \eqn{\mathrm{Sym}(p)}.
}
\note{
It is the inverse map of logarithm map by \code{\link[=euc.LogMap]{euc.LogMap()}}.
}
\examples{
p = 5
S1 = CovM(p)
S2 = CovM(p)
# Compute logarithm map at base point \eqn{S1} lying in tangent space \eqn{T_S\mathrm{Sym}(p)}
V = euc.LogMap(S1,S2)
euc.ExpMap(S1,V)
}
\seealso{
\code{\link[=euc.LogMap]{euc.LogMap()}}
}
