% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.geodesic}
\alias{euc.geodesic}
\title{Geodesic in symmetric matrix space}
\usage{
euc.geodesic(t, S1, S2)
}
\arguments{
\item{t}{Time. It should be in the interval of \eqn{t \in [0,1]}.}

\item{S1}{A symmetric matrix. It is the starting point of the geodesic \eqn{\gamma(t=0)} in  \eqn{\mathrm{Sym}(p)}.}

\item{S2}{A symmetric matrix. It is the endpoint of the geodesic \eqn{\gamma(t=1)} in  \eqn{\mathrm{Sym}(p)}.}
}
\value{
A symmetric matrix. It lies on the geodesic \eqn{\gamma(t)}  at time \eqn{t, t \in [0,1]}.
}
\description{
Compute geodesic \eqn{\gamma(t), t \in [0,1]} in symmetric matrix space  \eqn{\mathrm{Sym}(p)}.
}
\details{
It is defined by \eqn{\mathrm{Exp}_{S_1}(t V)}, where \eqn{V = \mathrm{Log}_{S_1}(S_2)} and \eqn{t \in [0,1]}.
}
\examples{
p = 5
S1 = CovM(p)
S2 = CovM(p)
euc.geodesic(t = 0.5,S1,S2)
}
