% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.inner_product}
\alias{euc.inner_product}
\title{Euclidean metric in symmetric matrix space}
\usage{
euc.inner_product(V1, V2)
}
\arguments{
\item{V1}{A tangent vector matrix.}

\item{V2}{A tangent vector matrix.}
}
\value{
A numerical distance.
}
\description{
Compute Euclidean metric for any two tangent vector matrices in symmetric matrix space \eqn{\mathrm{Sym}(p)}.
}
\details{
A Riemannian metric is defined at \eqn{S \in \mathrm{Sym}(p)} for any two tangent vectors \eqn{V,W \in T_S\mathrm{Sym}(p)= \mathrm{Sym}(p)}
as follows: \deqn{ g_S(V,W) = \mathrm{tr}(V^TW)}
This implies the inner product is in fact independent of the base point \eqn{S}.
}
\examples{
p = 5
S1 = CovM(p)
S2 = CovM(p)
S3 = CovM(p)
# Compute tangent vectors
V1 = S2-S1
V2 = S3-S1
euc.inner_product(V1,V2)
euc.inner_product(V2,V1)
}
