% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.LogMap}
\alias{euc.LogMap}
\title{Logarithm map in symmetric matrix space}
\usage{
euc.LogMap(S1, S2)
}
\arguments{
\item{S1}{A symmetric matrix.}

\item{S2}{A symmetric matrix.}
}
\value{
A tangent vector matrix in tangent space \eqn{T_{S_1}\mathrm{Sym}(p)}.
}
\description{
Logarithm map in symmetric matrix space \eqn{\mathrm{Sym}(p)}.
It aims to obtain a tangent vector matrix by computing the velocity from point \eqn{S_1} to point \eqn{S_2} both locating in \eqn{\mathrm{Sym}(p)}.
}
\note{
The tangent vector \eqn{V} also lies in symmetric matrix space \eqn{\mathrm{Sym}(p)}. It is the inverse map of exponential map by \code{\link[=euc.ExpMap]{euc.ExpMap()}}.
}
\examples{
p = 5
S1 = CovM(p)
S2 = CovM(p)
euc.LogMap(S1,S2)
}
\seealso{
\code{\link[=euc.ExpMap]{euc.ExpMap()}}
}
