% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.mean}
\alias{euc.mean}
\title{Euclidean mean in symmetric matrix space}
\usage{
euc.mean(S)
}
\arguments{
\item{S}{Array object with the dimension of \eqn{p \times p \times n}. It includes \eqn{n} observations of \eqn{p \times p} symmetric matrices.}
}
\value{
A symmetric matrix in \eqn{\mathrm{Sym}(p)}. It is the centre point of dataset \eqn{S}.
}
\description{
Compute Euclidean mean  in symmetric matrix space \eqn{\mathrm{Sym}(p)}
}
\examples{
S = lapply(1:15, function(i) CovM(5))
euc.mean(list2array(S))
}
