% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUCfuncs.R
\name{euc.ortho_basis}
\alias{euc.ortho_basis}
\title{Orthonormal basis at the base point of identity matrix in symmetric matrix space}
\usage{
euc.ortho_basis(p)
}
\arguments{
\item{p}{Row (or column) number of the symmetric matrix.}
}
\value{
Orthonormal basis in the tangent space \eqn{T_{I_p}\mathrm{Sym}(p)} in a list object with the dimension of \eqn{m= p(p+1)/2}.
}
\description{
Construct orthonormal basis in the tangent space \eqn{T_{I_p}\mathrm{Sym}(p)}.
}
\details{
Let \eqn{e_i} denote the \eqn{i}th standard basis vector in \eqn{\mathbb{R}^p},
where \eqn{e_i = (0,...,1,...,0)_p^T}. For any \eqn{1 \leq i \leq j \leq p},
we can construct an orthonormal basis in \eqn{T_{I_p}\mathrm{Sym}(p)} as
\deqn{ E_{ij} = e_ie_j^T \; \mathrm{when} \; i = j,}
and \deqn{ E_{ij} = \frac{\sqrt{2}}{2}\left(e_ie_j^T + e_je_i^T\right), \; \mathrm{when} \; i \neq j.}
To simplify the notation, let \eqn{E_{ij}} denote as \eqn{E_r} where \eqn{r = 1,...,m= p(p+1)/2}.
This represents the \eqn{r}th basis vector in the tangent space \eqn{T_{I_p}\mathrm{Sym}(p)}  at the point \eqn{ I_p \in \mathrm{Sym}(p)}.
}
\note{
According to the property of Euclidean geometry \eqn{\mathrm{Sym}(p)},
the orthonormal basis in tangent spaces does not depend on the base point ,
i.e., for given any \eqn{S \in \mathrm{Sym}(p)},
the orthonormal basis in the tangent space \eqn{T_{S}\mathrm{Sym}(p)} is exactly same as basis in the tangent space \eqn{T_{I_p}\mathrm{Sym}(p)} .
}
\examples{
euc.ortho_basis(p = 5)
}
