% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flattenlist}
\alias{flattenlist}
\title{Unified list}
\usage{
flattenlist(X)
}
\arguments{
\item{X}{List object with the length of \eqn{n}.}
}
\value{
A single list data.
}
\description{
Flatten the list data into a single list.
}
\examples{
X1 = lapply(1:5, function(i) CovM(5))
X2 = lapply(1:5, function(i) CovM(5))
X = list(X1=X1,X2 =X2)
Xlist = flattenlist(X)
Xlist
}
