% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{hor.ortho_basis}
\alias{hor.ortho_basis}
\title{Orthonormal basis in horizontal space}
\usage{
hor.ortho_basis(S)
}
\arguments{
\item{S}{A SPD matrix in \eqn{\mathcal{S}_p^+}.}
}
\value{
Orthonormal basis with the dimension of \eqn{m = p(p-1)/2}
}
\description{
Construct orthonormal basis in the horizontal space \eqn{\mathcal{H}_S}.
}
\note{
This functions only applies to the tangent space which is horizontal,
i.e. the vertical components in tangent space \eqn{T_S\mathcal{S}_p^+} are all 0.
}
\examples{
S = CovM(5)
hor.ortho_basis(S)
}
