% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{hor}
\alias{hor}
\title{Horizontal projection}
\usage{
hor(V, S)
}
\arguments{
\item{V}{A tangent vector matrix in \eqn{T_S\mathcal{S}_p^+}.}

\item{S}{A SPD matrix in \eqn{\mathcal{S}_p^+}.}
}
\value{
A symmetric matrix. It is the horizontal component of tangent vector \eqn{V \in T_S\mathcal{S}_p^+}.
}
\description{
The horizontal component \eqn{U \in \mathcal{H}_S} of the tangent vector \eqn{V \in T_S\mathcal{S}_p^+}.
}
\details{
Let \eqn{U \in \mathcal{H}_S } be a vertical component of tangent vector \eqn{V \in T_S\mathcal{S}_p^+}. We have
\deqn{d\pi_S(V) = d\pi_S(U) = X,}
where \eqn{X} is the projected tangent vector in the tangent space \eqn{T_{\pi(S)}\mathcal{C}_p^+}.
}
\note{
Notes could be found in notes of \code{\link[=ver]{ver()}}.
}
\examples{
S1 = CovM(5); S2 = CovM(5)
V1 = spd.LogMap(S1,S2)
hor(V1,S1)
}
\seealso{
\code{\link[=ver]{ver()}}
}
