% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.lower.tri}
\alias{is.lower.tri}
\title{Lower triangular matrix}
\usage{
is.lower.tri(M)
}
\arguments{
\item{M}{A square matrix.}
}
\value{
Logical value. Return \code{TRUE} if the given square matrix is a lower triangular matrix. Otherwise, output is \code{FALSE}.
}
\description{
Check whether a given squarematrix is lower triangular matrix or not.
}
\examples{
M = matrix(-c(1:9),3,3)
M[upper.tri(M,diag = FALSE)] = 0
is.lower.tri(M)
}
