% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCholeskySpace.R
\name{lch.ExpMap}
\alias{lch.ExpMap}
\title{Exponential map in \eqn{\mathcal{S}_p^+} with LogCholesky metric}
\usage{
lch.ExpMap(S, V)
}
\arguments{
\item{S}{Base point in manifold \eqn{\mathcal{S}_p^+}}

\item{V}{Tangent vector in tangent space \eqn{T_S\mathcal{S}_p^+}}
}
\value{
Mapped point in \eqn{\mathcal{S}_p^+}
}
\description{
Exponential map from \eqn{S} into direction \eqn{V}
}
