% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCholeskySpace.R
\name{lch.geodesic}
\alias{lch.geodesic}
\title{Geodesic in \eqn{\mathcal{S}_p^+} with LogCholesky metric}
\usage{
lch.geodesic(t, S1, S2)
}
\arguments{
\item{t}{Time which should be in the interval of \eqn{[0,1]}}

\item{S1}{Starting point in manifold \eqn{\mathcal{S}_p^+}}

\item{S2}{Ending point in manifold \eqn{\mathcal{S}_p^+}}
}
\value{
Data ponint the geodesic from \eqn{S_1} to \eqn{S_2}
}
\description{
Geodesic in \eqn{\mathcal{S}_p^+} with LogCholesky metric
}
