% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCholeskySpace.R
\name{lch.inner_product}
\alias{lch.inner_product}
\title{Inner product in \eqn{\mathcal{S}_p^+} with LogCholesky metric}
\usage{
lch.inner_product(X, Y, S)
}
\arguments{
\item{X}{Tangent vector in tangent space \eqn{T_S\mathcal{S}_p^+}}

\item{Y}{Tangent vector in tangent space \eqn{T_S\mathcal{S}_p^+}}

\item{S}{Base point}
}
\value{
A numerical value which measures the inner product between two tangent vectors
}
\description{
Inner product in \eqn{\mathcal{S}_p^+} with LogCholesky metric
}
