% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCholeskySpace.R
\name{lch.LogMap}
\alias{lch.LogMap}
\title{Logarithm map in \eqn{\mathcal{S}_p^+} with LogCholesky metric}
\usage{
lch.LogMap(S1, S2)
}
\arguments{
\item{S1}{Starting point in manifold \eqn{\mathcal{S}_p^+}}

\item{S2}{Ending point in manifold \eqn{\mathcal{S}_p^+}}
}
\value{
Tangent vector in \eqn{T_S\mathcal{S}_p^+}
}
\description{
Logarithm map emanating from \eqn{S_1} to \eqn{S_2}
}
