% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCholeskySpace.R
\name{lch.mean}
\alias{lch.mean}
\title{Parallel transport in \eqn{\mathcal{S}_p^+} with LogCholesky metric}
\usage{
lch.mean(S)
}
\arguments{
\item{S}{Array data}
}
\value{
Mean
}
\description{
Parallel transport in \eqn{\mathcal{S}_p^+} with LogCholesky metric
}
