% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCholeskySpace.R
\name{lch.metric}
\alias{lch.metric}
\title{Geodesic distance in \eqn{\mathcal{S}_p^+} with LogCholesky metric}
\usage{
lch.metric(S1, S2)
}
\arguments{
\item{S1}{Starting point in manifold \eqn{\mathcal{S}_p^+}}

\item{S2}{Ending point in manifold \eqn{\mathcal{S}_p^+}}
}
\value{
A numerical value of distance between \eqn{S_1} and \eqn{S_2} in \eqn{\mathcal{S}_p^+}
}
\description{
Geodesic distance in \eqn{\mathcal{S}_p^+} with LogCholesky metric
}
