% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogCholeskySpace.R
\name{lch.para_tran}
\alias{lch.para_tran}
\title{Parallel transport in \eqn{\mathcal{S}_p^+} with LogCholesky metric}
\usage{
lch.para_tran(S1, S2, X)
}
\arguments{
\item{S1}{Base point in \eqn{\mathcal{S}_p^+}}

\item{S2}{Objective point in \eqn{\mathcal{S}_p^+}}

\item{X}{Tangent vector in tangent space \eqn{T_{S_1}\mathcal{S}_p^+}}
}
\value{
Transported vector along the geodesic at \eqn{T_{S_2}\mathcal{S}_p^+}
}
\description{
Parallel transport in \eqn{\mathcal{S}_p^+} with LogCholesky metric
}
