% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2array}
\alias{list2array}
\title{List object to array object}
\usage{
list2array(X)
}
\arguments{
\item{X}{List object with length \eqn{n}. Note that each list contains a \eqn{p \times p} square matrix.}
}
\value{
An \eqn{p \times p \times n} array object.
}
\description{
Transform list object to array object.
}
\examples{
X = lapply(1:10, function(i) CovM(5))
S = list2array(X)

}
\seealso{
\code{\link[=array2list]{array2list()}}
}
