% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list2matrix}
\alias{list2matrix}
\title{List object to matrix object}
\usage{
list2matrix(X, byrow = TRUE)
}
\arguments{
\item{X}{List object with length \eqn{n}. Note that each list contains a \eqn{p \times p} square matrix.}

\item{byrow}{Logical values. When it is \code{TRUE}, we append list data by rows, vice versa.}
}
\value{
A stacked matrix with the dimension of \eqn{(n*p) \times p}
}
\description{
Transform list object to matrix object.
}
\examples{
X = lapply(1:10, function(i) CovM(5))
mat = list2matrix(X,byrow = TRUE)
mat
}
