% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MeanArray}
\alias{MeanArray}
\title{Mean of array data}
\usage{
MeanArray(M, dim = NULL)
}
\arguments{
\item{M}{An \eqn{p \times p \times n} array object.}

\item{dim}{A choice from \code{dim = (1,2,3)} corresponding to array object.
For example \code{dim = 3}, it means we compute array mean by the dimension of \code{3}.}
}
\value{
Matrix mean of dimension \code{dim}
}
\description{
Compute the mean of array data by the specific dimensions
}
\examples{
S = array(0,c(5,5,10))
for(i in 1:10){S[,,i] = CovM(5)}
mS1 = MeanArray(S,dim = 1)
mS2 = MeanArray(S,dim = 2)
mS3 = MeanArray(S,dim = 3)
}
