% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE.Diag}
\alias{MLE.Diag}
\title{MLE for the model with diagonal matrix coefficients}
\usage{
MLE.Diag(lagvec, maxlag = NULL, Mean = NULL)
}
\arguments{
\item{lagvec}{lagged vectors}

\item{maxlag}{maximum lags}

\item{Mean}{Logical values. True for involving Mean reversion, vice versa.}
}
\value{
estimated parameters and AIC values for different lags
}
\description{
MLE for the model with diagonal matrix coefficients
}
