% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE.Sca}
\alias{MLE.Sca}
\title{MLE for VAR model with scalar coefficients}
\usage{
MLE.Sca(lagvec, maxlag = NULL, Mean = NULL)
}
\arguments{
\item{lagvec}{lagged results obtained from functions \link{spd.coordvec_lags} or \link{cor.coordvec_lags}}

\item{maxlag}{maximum lags with default 10}

\item{Mean}{Logical values. True for involving Mean reversion, vice versa.}
}
\value{
estimated parameters and AIC values for different lags
}
\description{
MLE for VAR model with scalar coefficients
}
