% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{modselectSca}
\alias{modselectSca}
\title{Model selection for manifold-adapted model with scalar coefficients}
\usage{
modselectSca(v, lagv, theta, sigma)
}
\arguments{
\item{v}{A \eqn{n \times m} matrix with \eqn{m} dimensions and \eqn{n} observations.}

\item{lagv}{A list data with the length of \eqn{q} containing lagged vectors  that each of it has the same structure with \code{v}.}

\item{theta}{A vector of estimated regressive coefficients from the output of \code{\link[=EstScaTheta]{EstScaTheta()}}.}

\item{sigma}{Estimated standard deviation of the white noise from the output of function \code{\link[=EstScaSig]{EstScaSig()}}.}
}
\value{
AIC , BIC, and log-likelihood function values
}
\description{
Model selection for manifold-adapted model with scalar coefficients
}
\seealso{
\code{\link[=EstScaTheta]{EstScaTheta()}},\code{\link[=EstScaSig]{EstScaSig()}}.
}
