% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLS.R
\name{OLS.Sca}
\alias{OLS.Sca}
\title{OLS for model estimation}
\usage{
OLS.Sca(v, mv = NULL, lagv = NULL, intercept = NULL)
}
\arguments{
\item{v}{observed coordinated vectors}

\item{mv}{Mean vectors}

\item{lagv}{Lagged vectors}

\item{intercept}{Logical values. If it is TRUE, the model consider the drift}
}
\value{
the regression results which same as \code{\link[=lm]{lm()}}
}
\description{
OLS for model estimation
}
