% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{phi}
\alias{phi}
\title{Group action}
\usage{
phi(S, D)
}
\arguments{
\item{S}{A SPD matrix in \eqn{\mathcal{S}_p^+}.}

\item{D}{A diagonal matrix with positive entries in Lie group \eqn{\mathcal{D}_p^+}.}
}
\value{
A SPD matrix after group action, which is also in manifold \eqn{\mathcal{S}_p^+}.
}
\description{
Group action operation in SPD matrix space  \eqn{\mathcal{S}_p^+}.
}
\details{
The Lie group \eqn{\mathcal{D}_p^+} of diagonal matrices with positive entries acts
smoothly, properly, and freely on Riemannian manifold \eqn{\mathcal{S}_p^+} via the action \eqn{\phi} at  point \eqn{S \in \mathcal{S}_p^+}:
\deqn{ \mathcal{D}_p^+ \times \mathcal{S}_p^+  \rightarrow \mathcal{S}_p^+}
i.e., \eqn{\phi_D(S) = DSD}, where \eqn{D \in \mathcal{D}_p^+}
}
\examples{
S = CovM(5)
D = diag(c(1:5))
phi(S,D)
}
\references{
\insertRef{gallier2020differential}{geomTS}
}
