% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{print_Sca}
\alias{print_Sca}
\title{Outputs of model with scalar coefficients}
\usage{
print_Sca(model, fixed.lag = NULL)
}
\arguments{
\item{model}{results from function \code{\link[=MLE.Sca]{MLE.Sca()}}. The object could be either "MVAR" or "VAR"}

\item{fixed.lag}{choose the results by a fixed lag}
}
\value{
estimated parameters
}
\description{
Outputs of model with scalar coefficients
}
