% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{residMat.Diag}
\alias{residMat.Diag}
\title{Compute residual matrix for diagonal matrix coefficient}
\usage{
residMat.Diag(v, lagv, Theta)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}

\item{Theta}{estimated alpha}
}
\value{
residual matrix
}
\description{
Compute residual matrix for diagonal matrix coefficient
}
