% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{residMat.Sca}
\alias{residMat.Sca}
\title{Compute residual matrix for scalar coefficient model}
\usage{
residMat.Sca(v, lagv, theta)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}

\item{theta}{estimated alpha}
}
\value{
residual matrix
}
\description{
Compute residual matrix for scalar coefficient model
}
