% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sol.sqrtm}
\alias{sol.sqrtm}
\title{Inverse of the square root for a non-singular square matrix}
\usage{
sol.sqrtm(M)
}
\arguments{
\item{M}{A non-singular square matrix.}
}
\description{
Compute the inverse of the square root of a non-singular square matrix as \eqn{M^{-1/2}}
such that \eqn{MM = S^{-1}} and \eqn{S} is non-singular.
}
\examples{
M = CovM(5)
sol.sqrtm(M)
}
\references{
\insertRef{golub2013matrix}{geomTS}
}
\seealso{
\link{sqrtm}
}
