% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.coor_vec}
\alias{spd.coor_vec}
\title{Coordinate vector by vectorizing projection in affine invariant geometry}
\usage{
spd.coor_vec(V, S)
}
\arguments{
\item{V}{A tangent vector matrix in \eqn{T_S\mathcal{S}_p^+} .}

\item{S}{A SPD matrix.}
}
\value{
A coordinate vector in \eqn{\mathbb{R}^m}.
}
\description{
Vectorize a tangent vector matrix \eqn{V \in T_S\mathcal{S}_p^+} into a coordinate vector \eqn{v \in \mathbb{R}^m}. Thereby,
the dimension of affine invariant geometry \eqn{\mathcal{S}_p^+} is \eqn{m = p(p+1)/2}.
}
\details{
Given any tangent vector \eqn{V \in T_S\mathcal{S}_p^+},
we can calculate its orthonormal coordinate vector \eqn{\boldsymbol{v} = (v_1,...,v_m)^T} in \eqn{\mathbb{R}^m} as
\deqn{\boldsymbol{v} = \langle V,S^{1/2}\boldsymbol{E}S^{1/2}  \rangle_S = \mathrm{tr}( S^{-1/2}VS^{-1/2}\boldsymbol{E}))}
where \eqn{\boldsymbol{E} = (E_1,\ldots,E_m)} is the orthonormal basis in \eqn{I_p} as shown in \link{spd.ortho_basis}.
Thereby, vectorization of \eqn{V \in T_S\mathcal{S}_p^+}  is mapped to \eqn{\boldsymbol{v} = (v_1,\ldots,v_m)^T \in \mathbb{R}^m}.
}
\examples{
S1 = CovM(3)
S2 = CovM(3)
V1 = spd.LogMap(S1,S2)
spd.coor_vec(V1,S1)
}
\seealso{
\code{\link[=spd.Euc2Manif]{spd.Euc2Manif()}}
}
