% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{spd.coordvec_lags}
\alias{spd.coordvec_lags}
\title{Coordinate vectors with lags in affine invariant geometry}
\usage{
spd.coordvec_lags(S, FSM = NULL, maxlag = NULL)
}
\arguments{
\item{S}{An \eqn{p \times p \times n} array data consisting of SPD matrices.}

\item{FSM}{A SPD matrix. It could be the Frechet sample mean \eqn{S} or an attract point.
If this value is missing, the mean reversion coordinate vector will not be considered in the model.}

\item{maxlag}{Maximum lags. The default value is 10.}
}
\value{
A set of coordinate vectors, including lagged vectors, observed vectors,
mean reversion vectors (If FSM is not \code{NULL}), and original matrices.
}
\description{
Compute coordinate vectors in  \eqn{\mathbb{R}^{p(p+1)/2}} including observed tangent vectors,
mean reversion vectors and lagged vectors in affine invariant geometry \eqn{\mathcal{S}_p^+}.
}
\seealso{
\code{\link[=spd.meanvecs]{spd.meanvecs()}},\code{\link[=spd.paraV]{spd.paraV()}},\code{\link[=spd.lagvecs]{spd.lagvecs()}}
}
