% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDSplot.R
\name{spd.dist}
\alias{spd.dist}
\title{Distance matrix using affine invariant metric}
\usage{
spd.dist(S, parallel = NULL, ncores = NULL)
}
\arguments{
\item{S}{An \eqn{p \times p \times n} array data consisting of SPD matrices.}

\item{parallel}{Logical value with default \code{FALSE}.When it is \code{TRUE},
we use parallel computation to compute distance matrix.}

\item{ncores}{The number of cores used in the parallel computation.}
}
\value{
Distance object as \code{\link[=dist]{dist()}}.
}
\description{
Compute distance matrix of SPD matrices in affine invariant geometry \eqn{\mathcal{S}_p^+}.
}
\examples{
data(EEG)
redEEG = DR_PCA(EEG$seizure1,N0 = 15)
spd.dist(redEEG)
}
\seealso{
\code{\link[=spd.metric]{spd.metric()}}
}
