% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.Euc2Manif}
\alias{spd.Euc2Manif}
\title{Mapping back the coordinate vector to data point in affine invariant geometry}
\usage{
spd.Euc2Manif(v, S)
}
\arguments{
\item{v}{Coordinate vector in \eqn{\mathbb{R}^m}.}

\item{S}{A SPD matrix.}
}
\value{
A SPD matrix in \eqn{\mathcal{S}_p^+}
}
\description{
Map back the coordinate vector \eqn{\boldsymbol{v} \in \mathbb{R}^m} to the tangent space \eqn{T_S\mathcal{S}_p^+}
and then do exponential map the base point \eqn{S \in \mathcal{S}_p^+}.
}
\details{
It follows the following process:
\deqn{V = \sum_{r = 1}^{m}v_rS^{1/2}E_rS^{1/2}\; \text{and then} \; S^* = \mathrm{Exp}_{S}(V).}
}
\examples{
S1 = CovM(5); S2 = CovM(5)
V = spd.LogMap(S1,S2)
v = spd.coor_vec(V,S1)
spd.Euc2Manif(v,S1)
}
\seealso{
\code{\link[=spd.coor_vec]{spd.coor_vec()}}
}
