% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.ExpMap}
\alias{spd.ExpMap}
\title{Exponential map in affine invariant geometry}
\usage{
spd.ExpMap(S, V)
}
\arguments{
\item{S}{A SPD matrix.}

\item{V}{A tangent vector matrix in \eqn{T_{S}\mathcal{S}_p^+}.}
}
\value{
A SPD matrix. It is the mapped point of \eqn{V} at base point \eqn{S} in the manifold \eqn{\mathcal{S}_p^+}.
}
\description{
Exponential map in affine invariant geometry \eqn{\mathcal{S}_p^+}.
It projects the tangent vector matrix \eqn{V \in T_{S}\mathcal{S}_p^+} at point \eqn{S \in \mathcal{S}_p^+} back to \eqn{\mathcal{S}_p^+}.
}
\note{
It is the inverse map of \code{\link[=spd.LogMap]{spd.LogMap()}}.
}
\examples{
S1 = CovM(3)
S2 = CovM(3)
V1 = spd.LogMap(S1,S2)
spd.ExpMap(S1,V1)
}
\seealso{
\code{\link[=spd.LogMap]{spd.LogMap()}}
}
