% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.geodesic}
\alias{spd.geodesic}
\title{Geodesic in affine invariant geometry}
\usage{
spd.geodesic(t, S1, S2)
}
\arguments{
\item{t}{Time. It should be in the interval of \eqn{[0,1]}.}

\item{S1}{A SPD matrix. It is the starting point of the geodesic  \eqn{\gamma(t =0)} in \eqn{\mathcal{S}_p^+}.}

\item{S2}{A SPD matrix. It is the endpoint of the geodesic \eqn{\gamma(t =1)} in \eqn{\mathcal{S}_p^+}.}
}
\value{
A SPD matrix. It lies on the geodesic \eqn{\gamma(t)} at time \eqn{t, t\in[0,1]}.
}
\description{
Compute geodesic \eqn{\gamma(t), t \in [0,1]} in affini invariant geometry \eqn{\mathcal{S}_p^+}.
}
\details{
It is defined by \eqn{\mathrm{Exp}_{S_1}(t V)}, where \eqn{V = \mathrm{Log}_{S_1}(S_2)} and \eqn{t \in [0,1]}.
}
\examples{
S1= CovM(3)
S2 = CovM(3)
spd.geodesic(t = 0.5,S1,S2)
spd.geodesic(t = 0.1,S1,S2)
}
