% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.inner_product}
\alias{spd.inner_product}
\title{Inner product in affine invariant geometry}
\usage{
spd.inner_product(V1, V2, S)
}
\arguments{
\item{V1}{A tangent vector matrix in  \eqn{T_S\mathcal{S}_p^+}.}

\item{V2}{A tangent vector matrix in  \eqn{T_S\mathcal{S}_p^+}.}

\item{S}{A SPD matrix. It has a tangent space \eqn{T_S\mathcal{S}_p^+}.}
}
\value{
A numerical distance.
}
\description{
Compute the inner product for any two tangent vector matrices \eqn{V_1, V_2 \in T_S\mathcal{S}_p^+} at the base point \eqn{S \in \mathcal{S}_p^+ }.
}
\details{
For any tangent vectors \eqn{V_1, V_2} in the tangent space \eqn{T_S\mathcal{S}_p^+} at the base point of \eqn{S \in \mathcal{S}_p^+},
the inner product of the Riemannian metric is defined as
\deqn{ \langle V_1,W_2 \rangle_S =  \mathrm{tr}(S^{-1}V_1S^{-1}W_2).}
}
\note{
Two tangent vectors should both lie in the same tangent space \eqn{T_S\mathcal{S}_p^+}.
}
\examples{
S1 = CovM(3)
S2 = CovM(3)
S3 = CovM(3)
V1 = spd.LogMap(S1,S2)
V2 = spd.LogMap(S1,S3)
spd.inner_product(V1,V2,S1)
}
