% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{spd.lagvecs}
\alias{spd.lagvecs}
\title{Lagged coordinate vectors from parallel transport in affine invariant geometry}
\usage{
spd.lagvecs(q, S, paraV)
}
\arguments{
\item{q}{A numerical value of lags.}

\item{S}{An (\eqn{p \times p \times n}) array data consisting of SPD matrices.}

\item{paraV}{Parallel transported tangent vectors from the output of \code{\link[=spd.paraV]{spd.paraV()}}.}
}
\value{
Lagged \eqn{q} coordinate vectors with \eqn{q} lists.
Each list is a matrix with the row of translated coordinate vectors and the column of dimensions of coordinate vectors.
}
\description{
Compute lagged coordinate vectors from parallel transport and arrange the structure of coordinate vectors.
}
\seealso{
\code{\link[=spd.coordvec_lags]{spd.coordvec_lags()}},\code{\link[=spd.paraV]{spd.paraV()}}
}
