% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.LogMap}
\alias{spd.LogMap}
\title{Logarithm map in affine invariant geometry}
\usage{
spd.LogMap(S1, S2)
}
\arguments{
\item{S1}{A SPD matrix.}

\item{S2}{A SPD matrix.}
}
\value{
A tangent vector matrix in \eqn{T_{S}\mathcal{S}_p^+}.
}
\description{
Logarithm map in affine invariant geometry \eqn{\mathcal{S}_p^+}.
It aims to obtain a tangent vector matrix by computing the velocity from point \eqn{S_1} to point \eqn{S_2} both locating in \eqn{\mathcal{S}_p^+}.
}
\note{
\itemize{
\item \eqn{T_{S}\mathcal{S}_p^+} is equivalent to the symmetric matrix space \eqn{Sym(p)}.
\item The tangent vector is the initial velocity at the starting point in the geodesic connecting \eqn{S_1} and \eqn{S_2} both in\eqn{\mathcal{S}_p^+}.
\item It is the inverse map of \code{\link[=spd.ExpMap]{spd.ExpMap()}}.
}
}
\examples{
S1 = CovM(3);S2 = CovM(3)
spd.LogMap(S1,S2)
}
\seealso{
\code{\link[=spd.ExpMap]{spd.ExpMap()}}
}
