% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{spd.meanvecs}
\alias{spd.meanvecs}
\title{Coordinate vectors of mean reversion in affine invariant geometry}
\usage{
spd.meanvecs(S, FSM)
}
\arguments{
\item{S}{An \eqn{p \times p \times n} array data consisting of SPD matrices.}

\item{FSM}{A SPD matrix. It could be the Frechet sample mean of data \eqn{S} or a given attract point.}
}
\value{
A matrix containing coordinate vectors of mean reversion with the row of observations and the column of the dimension \eqn{m}.
}
\description{
Compute coordinate vectors at the centre point or attract point in affine invariant geometry \eqn{\mathcal{S}_p^+}.
}
\details{
Let \eqn{\tilde{S}} be Frechet sample mean or attract point. We can compute the tangent vectors for all data points into tangent space \eqn{T_{\tilde{S}}\mathcal{S}_p^+} by
\deqn{V_i = \mathrm{Log}_{\tilde{S}}(S_i), i = 1,\ldots,n.}
Subsequently, we have the coordinate vectors by vectorization \eqn{Vec(V_i)} by \code{\link[=spd.coor_vec]{spd.coor_vec()}}.
}
\seealso{
\code{\link[=spd.coordvec_lags]{spd.coordvec_lags()}}
}
