% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.metric}
\alias{spd.metric}
\title{Geodesic distance in affine invariant geometry}
\usage{
spd.metric(S1, S2)
}
\arguments{
\item{S1}{A SPD matrix.}

\item{S2}{A SPD matrix.}
}
\value{
A numerical value of geodesic distance between \eqn{S_1} and \eqn{S_2} in \eqn{\mathcal{S}_p^+}
}
\description{
Calculate geodesic distance for any two SPD matrices \eqn{S_1,S_2} in SPD matrix space \eqn{\mathcal{S}_p^+}.
}
\details{
Given any two points \eqn{S_1} and \eqn{S_2}  in SPD matrix space \eqn{\mathcal{S}_p^+}, we can define the Riemannian affine invariant metric as
\deqn{d(S_1,S_2) = \sqrt{\sum_{i=1}^{p}\log^2(\lambda_i)}}
where \eqn{\lambda_i,i = 1,\ldots,p} are eigenvalues of matrix \eqn{S_1^{-1/2}S_2S_1^{-1/2}}.
}
\note{
Note that \eqn{d(S_1,S_2) = d(S_2,S_1)}
}
\examples{
S1 = CovM(3)
S2 = CovM(3)
spd.metric(S1,S2)
spd.metric(S2,S1)
}
