% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.ortho_basis}
\alias{spd.ortho_basis}
\title{Orthonormal basis at any base point in affine invariant geometry}
\usage{
spd.ortho_basis(S)
}
\arguments{
\item{S}{A SPD matrix.}
}
\value{
A list of orthonormal basis in the tangent space \eqn{T_S\mathcal{S}_p^+}
containing \eqn{m = \frac{1}{2}p(p+1)} observations of \eqn{p \times p} basis.
}
\description{
Construct orthonormal basis in the tangent space \eqn{T_S\mathcal{S}_p^+}.
}
\details{
An orthonormal basis of \eqn{T_S\mathcal{S}_p^+} for each  \eqn{S \in \mathcal{S}_p^+} can be defined
via parallel transport of a choice of orthonormal basis of the tangent space at the identity matrix \eqn{I_p \in T_{I_p}\mathcal{S}_p^+} .
The matrices \eqn{E_r, r= 1,\ldots,m = p(p+1)/2}, defined in \code{\link[=euc.coor_vec]{euc.coor_vec()}} determine an orthonormal basis of the tangent space at
\eqn{I_p} with respect to the Riemannian inner product. Using parallel transport from \eqn{I_p} to a general point \eqn{S \in \mathcal{S}_p^+} in \code{\link[=spd.para_tran]{spd.para_tran()}},
we obtain an orthonormal basis of \eqn{T_S\mathcal{S}_p^+} defined by
\deqn{ E^{S}_r = \mathcal{P}_{I_p \rightarrow S}(E_r) = S^{1/2}E_rS^{1/2}. }
This basis can be thought of as a global smooth section of the frame bundle of the Riemannian manifold \eqn{\mathcal{S}_p^+}.
}
\note{
By definition, it follows that affine invariant geometry is parallelizable.
}
\examples{
spd.ortho_basis(diag(3))
S1 = CovM(3)
spd.ortho_basis(S1)
}
