% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIfuncs.R
\name{spd.para_tran}
\alias{spd.para_tran}
\title{Parallel transport in affine invariant geometry}
\usage{
spd.para_tran(S1, S2, V)
}
\arguments{
\item{S1}{A SPD matrix.}

\item{S2}{A SPD matrix.}

\item{V}{A tangent vector matrix in \eqn{T_{S_1}\mathcal{S}_p^+}.}
}
\value{
A translated tangent vector matrix in \eqn{T_{S_2}\mathcal{S}_p^+}.
}
\description{
Parallel transport the tangent vector \eqn{V \in T_{S_1}\mathcal{S}_p^+}
into another tangent space \eqn{T_{S_2}\mathcal{S}_p^+} along the geodesic connecting \eqn{S_1} and \eqn{S_2} in \eqn{\mathcal{S}_p^+}.
}
\details{
Parallel transport involves the translation of a tangent vector \eqn{V} from the tangent space \eqn{T_{S_1}\mathcal{S}_p^+}
at point \eqn{S_1 \in \mathcal{S}_p^+} to the tangent space \eqn{T_{S_2}\mathcal{S}_p^+} at point \eqn{S_2 \in \mathcal{S}_p^+}
and its  mathematical formula is :
\deqn{\mathcal{P}_{S_1\rightarrow S_2}(V) = WV_1W^T.}
Here, \eqn{W} can be simply written  as  \eqn{W = (S_2S_1^{-1})^{1/2}}.
}
\examples{
S1 = CovM(3)
S2 = CovM(3)
V1 = spd.LogMap(S1,S2)
spd.para_tran(S1,S2,V1)
}
\references{
\insertRef{yair2019parallel}{geomTS}
}
