% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagVecs.R
\name{spd.paraV}
\alias{spd.paraV}
\title{Parallel transport vectors with maximum lags in affine invariant geometry}
\usage{
spd.paraV(maxlag, S, V)
}
\arguments{
\item{maxlag}{A numerical value which represents the maximum lags involved}

\item{S}{An (\eqn{p \times p \times n}) array data consisting of SPD matrices.}

\item{V}{A (\eqn{n-1}) length list data consisting of tangent vectors}
}
\value{
A list object for the parallel transported coordinate vectors
}
\description{
Compute lagged tagent vector matrix by parallel transport in affine invariant geometry \eqn{\mathcal{S}_p^+}.
}
\details{
Given any original data point \eqn{S} in \eqn{\mathcal{S}_p^+} and corresponding
tangent vectors \eqn{V}, we compute all parallel transport vectors with the maximum lags \eqn{maxlag}.
For example, \code{maxlag = 3}, the parallel transported vector could be written as
\eqn{V_{4,1} = \mathcal{P}_{S_1,S_4}(V_1)}, \eqn{V_{4,2} = \mathcal{P}_{S_2,S_4}(V_2)}, and \eqn{V_{4,3} = \mathcal{P}_{S_3,S_4}(V_3)}.
The basic structure (e.g maxlag = 3) stored in outputs is

\eqn{\mathcal{P}_{S_1,S_2}(V_1)};

\eqn{\mathcal{P}_{S_2,S_3}(V_2)}, \eqn{\mathcal{P}_{S_1,S_3}(V_1)};

\eqn{\mathcal{P}_{S_3,S_4}(V_3)}, \eqn{\mathcal{P}_{S_2,S_4}(V_2)} ,\eqn{\mathcal{P}_{S_1,S_4}(V_1)};

\eqn{\mathcal{P}_{S_4,S_5}(V_4)}, \eqn{\mathcal{P}_{S_3,S_5}(V_3)} ,\eqn{\mathcal{P}_{S_2,S_5}(V_2)}.
}
\seealso{
\code{\link[=spd.lagvecs]{spd.lagvecs()}}
}
