% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmodel.R
\name{squaredR}
\alias{squaredR}
\title{\eqn{R^2} for checking the goodness of fit}
\usage{
squaredR(v, lagvs, theta)
}
\arguments{
\item{v}{Coordinate vectors of observed data.}

\item{lagvs}{Coordinate vectors of lagged vectors.}

\item{theta}{Estimated parameters. If it is a list, the model with diagonal matrix coefficients is used.
Otherwise, it is the model with scalar coefficients.}
}
\value{
\eqn{R^2}
}
\description{
Compute \eqn{R^2} for evaluating the goodness of model fitting.
}
