% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{submer}
\alias{submer}
\title{Riemannian submersion}
\usage{
submer(S)
}
\arguments{
\item{S}{A SPD matrix in \eqn{\mathcal{S}_p^+}.}
}
\value{
A SPD matrix in quotient manifold \eqn{\mathcal{C}_p^+} which has 1 entries for all diagonals.
}
\description{
Riemannian submersion \eqn{\pi} between manifold \eqn{\mathcal{S}_p^+} and manifold \eqn{\mathcal{C}_p^+}.
}
\details{
Let \eqn{\cdot :  \mathcal{D}_p^+ \times \mathcal{S}_p^+ \rightarrow \mathcal{S}_p^+} be a
smooth, free and proper action, with \eqn{ \mathcal{S}_p^+} a Lie group acting by isometries of \eqn{ \mathcal{S}_p^+}.
The smooth manifold \eqn{ \mathcal{C}_p^+ =  \mathcal{S}_p^+/ \mathcal{D}_p^+} can be seen as the quotient manifold via Riemannian submersion
\eqn{\pi:  \mathcal{S}_p^+ \rightarrow  \mathcal{C}_p^+}.
For example, a full rank correlation matrix from a covariance matrix is obtained by
\deqn{\pi : S \in \mathcal{S}_p^+ \rightarrow C =  \Delta_S^{-1}S\Delta_S^{-1} \in \mathcal{C}_p^+.}
where \eqn{\Delta_S = \sqrt{\mathrm{Diag}(S)}} and \eqn{\mathrm{Diag}(S) \in \mathcal{D}_p^+} is the diagonal matrix of \eqn{S}.
}
\note{
In this package, we consider quotient manifold consisting of full-rank correlation matrices.
Therefore, correlation matrix \code{\link[=submer]{submer()}} also could be achieved by function \link{cov2cor}.
}
\examples{
S = CovM(5)
submer(S)
}
\references{
\insertRef{david2019riemannian}{geomTS}
}
