% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tr}
\alias{tr}
\title{Trace of a square matrix}
\usage{
tr(M)
}
\arguments{
\item{M}{A square matrix.}
}
\value{
A numerical value. It is the sum of diagonals for a given square matrix \eqn{M}
}
\description{
Compute the trace of a square matrix \eqn{M}.
}
\examples{
tr(matrix(1:9,3,3))
}
