% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QIfuncs.R
\name{ver}
\alias{ver}
\title{Vertical projection}
\usage{
ver(V, S)
}
\arguments{
\item{V}{A tangent vector matrix in \eqn{T_S\mathcal{S}_p^+}.}

\item{S}{A SPD matrix in  \eqn{\mathcal{S}_p^+}.}
}
\value{
A symmetric matrix. It is the vertical component of tangent vector \eqn{V \in T_S\mathcal{S}_p^+}.
}
\description{
The vertical component \eqn{W \in \mathcal{V}_S} of the tangent vector \eqn{V \in T_S\mathcal{S}_p^+}.
}
\details{
Let \eqn{W \in \mathcal{V}_S} be the vertical component of the tangent vector \eqn{V \in T_S\mathcal{S}_p^+}. We have
\deqn{d\pi_S(W) = 0.}
}
\note{
The tangent sapce \eqn{T_{S}\mathcal{S_p^+}} to \eqn{\mathcal{S_p^+}} at \eqn{S} splits into the two components
\deqn{T_S\mathcal{S_p^+} =  \mathcal{H}_S \oplus  \mathcal{V}_S.}

Let \eqn{W \in \mathcal{V}_S} be the vertical component of the tangent vector \eqn{V \in T_S\mathcal{S}_p^+}, and
\eqn{U \in \mathcal{H}_S} be the horizontal component of the tangent vector \eqn{V \in T_S\mathcal{S}_p^+}. We have
\deqn{V = U + W,}
and
\deqn{d\pi_S(W) = 0; d\pi_S(U) = d\pi_S(V) .}
This illustrates \eqn{d\pi_S} gives a linear isomorphism  between \eqn{\mathcal{H}_S} and \eqn{T_{\pi(S)}\mathcal{C}_p^+}.
}
\examples{
S1 = CovM(5)
S2 = CovM(5)
V1 = spd.LogMap(S1,S2)
ver(V1,S1)
}
\seealso{
\code{\link[=hor]{hor()}}
}
