% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.utils.R
\name{vertical.image.legend}
\alias{vertical.image.legend}
\title{Put color bar legend on the right side of an existing plot.}
\usage{
vertical.image.legend(zlim, col)
}
\arguments{
\item{zlim}{charterers with initial index and end index}

\item{col}{colour bar}
}
\description{
Put color bar legend on the right side of an existing plot.
}
