% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wMLE.R
\name{wMLE.Sca}
\alias{wMLE.Sca}
\title{Weighted MLE}
\usage{
wMLE.Sca(LAGVEC, maxlag = NULL, Mean = NULL)
}
\arguments{
\item{LAGVEC}{Lagged coordinate vectors from the outputs of \code{\link[=euc.coordvec_lags]{euc.coordvec_lags()}}, \code{\link[=spd.coordvec_lags]{spd.coordvec_lags()}}, and \code{\link[=cor.coordvec_lags]{cor.coordvec_lags()}}}

\item{maxlag}{A numerical value of maximum lag}

\item{Mean}{Logical value with default \code{FALSE} that the regression model does not consider mean reversion term.}
}
\value{
A "MVAR" or "VAR" object containing \describe{
\item{theta}{Estimated regression coefficients}
\item{sigma}{Estimated s.d.}
\item{AIC}{AIC values}
\item{liglik}{log-likelihood values}
}
}
\description{
Weighted MLE
}
