% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrapGau}
\alias{wrapGau}
\title{Wrapped Gaussian distribution}
\usage{
wrapGau(S, sig)
}
\arguments{
\item{S}{A SPD matrix}

\item{sig}{A positive value. It is the standard deviation of normal distribution with zero mean.}
}
\value{
A SPD matrix
}
\description{
Generate a symmetric positive definite matrix from a wrapped Gaussian distribution.
}
\details{
The wrapped Gaussian distribution represents a bump of density around the point \eqn{S} with the degree of spread and shape of the density contours determined by the matrix \eqn{\Sigma}.
In fact we will work with a related set of distributions obtained by adding a non-zero mean \eqn{w \in \mathbb{R}^m} to the Gaussian distribution. In this case we write
\deqn{U \sim N_{T_S\mathcal{M}_p} (w,\Sigma)}
to denote the distribution of the random tangent vector \eqn{U}. The resulting distribution of the random point
\deqn{X= \mathrm{Exp}_{S}(U)}
closely approximates the wrapped Gaussian with zero mean centred at the point \eqn{\mathrm{Exp}_{S} (\sum_i w_iv_i)}
when the vector \eqn{w} is sufficiently small, and \eqn{v_i} is the \eqn{i}th orthonormal basis in \eqn{T_{S}\mathcal{M}_p} and \eqn{i = 1,\ldots,m}.
}
\note{
\eqn{\Sigma = \sigma^2 I_m}.
}
\examples{
wrapGau(diag(5),0.2)
}
\references{
\insertRef{mallasto2018wrapped}{geomTS}
}
\seealso{
\code{\link[=CovM]{CovM()}}
}
