% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstDiag}
\alias{EstDiag}
\title{Estimate Parameters in diagonal matrix VAR model}
\usage{
EstDiag(v, lagv)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}
}
\description{
Estimate Parameters in diagonal matrix VAR model
}
