% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstDiagSig}
\alias{EstDiagSig}
\title{Estimate diagonals of Sigma in VAR model}
\usage{
EstDiagSig(v, lagv, Theta)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}

\item{Theta}{estimated diagonals of coefficients in VAR models}
}
\value{
estimated diagonal of Sigma
}
\description{
Estimate diagonals of Sigma in VAR model
}
