% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstDiagTheta}
\alias{EstDiagTheta}
\title{Regressive coefficients in the manifold-adapted model with diagonal coefficients}
\usage{
EstDiagTheta(v, lagv)
}
\arguments{
\item{v}{A \eqn{n \times m} matrix with \eqn{m} dimensions and \eqn{n} observations.}

\item{lagv}{A list data with the length of \eqn{q} containing lagged vectors  that each of it has the same structure with \code{v}.}
}
\value{
A list of estimated diagonal regressive coefficients with the length of \eqn{q}.
}
\description{
Regressive coefficients in the manifold-adapted model with diagonal coefficients
}
