% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstSca}
\alias{EstSca}
\title{Estimate Parameters in the model with scalar coefficients}
\usage{
EstSca(v, lagv)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}
}
\value{
estimated scalar coefficients
}
\description{
Estimate Parameters in the model with scalar coefficients
}
\seealso{
The whole function could find in function \code{\link[=MLE.Sca]{MLE.Sca()}}
}
