% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{EstScaTheta}
\alias{EstScaTheta}
\title{Estimate alpha}
\usage{
EstScaTheta(v, lagv)
}
\arguments{
\item{v}{vectors}

\item{lagv}{lagged vectors}
}
\value{
estimated alpha
}
\description{
Estimate alpha
}
\seealso{
The whole function could find in function \code{\link[=MLE.Sca]{MLE.Sca()}} and \code{\link[=EstSca]{EstSca()}}
}
