% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE.Diag}
\alias{MLE.Diag}
\title{MLE for manifold-adapted model with diagonal coefficients}
\usage{
MLE.Diag(lagvec, maxlag = NULL, Mean = NULL)
}
\arguments{
\item{lagvec}{Lagged vector results obtained from functions \link{euc.coordvec_lags}, \link{spd.coordvec_lags} nor \link{cor.coordvec_lags}.}

\item{maxlag}{Maximum lag with default value of \code{10}.}

\item{Mean}{Logical value with default value of \code{FALSE}. When it is \code{TRUE}, we will consider the mean reverting term
in our manifold-adapted time series model, vice versa.}
}
\value{
Estimated parameters and model parameters, including \describe{
\item{Theta}{A list of estimated diagonal coefficients of autoregressive terms and mean reverting term (if \code{Mean} is \code{TRUE}).}
\item{Sigma}{A vector of estimated diagonals of standard deviation for the white noise}
\item{AIC}{AIC values}
\item{BIC}{BIC values}
\item{loglik}{log-likelihood function value.}
}
}
\description{
Model inference by MLE for manifold-adapted model with diagonal coefficients.
}
