% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{MLE.Sca}
\alias{MLE.Sca}
\title{MLE for manifold-adapted model with scalar coefficients}
\usage{
MLE.Sca(lagvec, maxlag = NULL, Mean = NULL)
}
\arguments{
\item{lagvec}{Lagged vector results obtained from functions \link{euc.coordvec_lags}, \link{spd.coordvec_lags} nor \link{cor.coordvec_lags}.}

\item{maxlag}{Maximum lag with default value of \code{10}.}

\item{Mean}{Logical value with default value of \code{FALSE}. When it is \code{TRUE}, we will consider the mean reverting term
in our manifold-adapted time series model, vice versa.}
}
\value{
Estimated parameters and model parameters, including \describe{
\item{theta}{Estimated parameters of autoregressive terms and mean reverting term (if \code{Mean} is \code{TRUE}).}
\item{sigma}{Estimated standard deviation of the white noise in the model.}
\item{AIC}{AIC value.}
\item{loglik}{log-likelihood function value.}
}
}
\description{
Model inference by MLE for manifold-adapted model with scalar coefficients.
}
\seealso{
\link{euc.coordvec_lags}, \link{spd.coordvec_lags}, \link{cor.coordvec_lags}.
}
