% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{print_Sca}
\alias{print_Sca}
\title{Print model coefficients in manifold-adapted model with scalar coefficients}
\usage{
print_Sca(model, fixed.lag = NULL)
}
\arguments{
\item{model}{Results from function \code{\link[=MLE.Sca]{MLE.Sca()}}. It is the object of \code{VAR} for the model with autoregressive terms only
or \code{MVAR} for the model with autoregressive and mean-reverting term.}

\item{fixed.lag}{Fixed lag. If it is \code{NULL}, the AIC method is used to suggest model lags.}
}
\value{
A vector of estimated parameters.
}
\description{
Print model coefficients in manifold-adapted model with scalar coefficients
}
\seealso{
\code{\link[=MLE.Sca]{MLE.Sca()}}
}
