% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLE.R
\name{residMat.Sca}
\alias{residMat.Sca}
\title{Residual matrix after model fitting with scalar coefficient}
\usage{
residMat.Sca(v, lagv, theta)
}
\arguments{
\item{v}{A \eqn{n \times m} matrix with \eqn{m} dimensions and \eqn{n} observations.}

\item{lagv}{A list data with the length of \eqn{q} containing lagged vectors  that each of it has the same structure with \code{v}.}

\item{theta}{A vector of estimated regressive coefficients from the output of \code{\link[=EstScaTheta]{EstScaTheta()}}.}
}
\value{
Residual matrix which has the same dimension with  \code{v}.
}
\description{
Residual matrix after model fitting with scalar coefficient
}
\seealso{
\code{\link[=EstScaTheta]{EstScaTheta()}}.
}
