% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spd.simulate}
\alias{spd.simulate}
\title{Simulate SPD matrices under the framework of manifold-adapted time series model with scalar coefficients}
\usage{
spd.simulate(n, alpha, sigma, beta = NULL, attS = NULL, p = NULL)
}
\arguments{
\item{n}{Number of total simulated time points.}

\item{alpha}{A vector. It represents the autoregressive coefficients.}

\item{sigma}{Standard deviation of white noise}

\item{beta}{The regression coefficient of mean reverting term. It should be given when \code{attS} is not \code{NULL}.}

\item{attS}{An attract point in \eqn{\mathcal{S}_p^+}. It should be given when \code{beta} is not \code{NULL}.}

\item{p}{The dimension \eqn{p}, if \code{beta} and \code{attS} are missing.}
}
\value{
An  \eqn{p \times p \times n} array data in \eqn{\mathcal{S}_p^+}.
}
\description{
Simulating SPD matrices in affine invariant manifold \eqn{\mathcal{S}_p^+} according to the mean-reverting autoregressive model
with scalar coefficients.
}
