% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{spd.simulate}
\alias{spd.simulate}
\title{Simulate SPD matrices}
\usage{
spd.simulate(n, alpha, sigma, beta = NULL, attS = NULL, p = NULL)
}
\arguments{
\item{n}{Number of time points}

\item{alpha}{Coefficients in VAR model whose length is the lags of VAR terms}

\item{sigma}{Standard deviation of white noise}

\item{beta}{Coefficient of mean reversion. It follows the attract point/ Mean}

\item{attS}{Attract point.}

\item{p}{Dimensions if \code{beta} and \code{attS} are missing.}
}
\value{
A array data \eqn{p \times p \times n}
}
\description{
Simulate SPD matrices in \eqn{\mathcal{S}_p^+}
}
